+++ CaesarCipher_ArithmeticEval/src/main/java/org/example/Main.java
@@ -8,8 +8,7 @@ import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
 
-//TIP To <b>Run</b> code, press <shortcut actionId="Run"/> or
-// click the <icon src="AllIcons.Actions.Execute"/> icon in the gutter.
+
 public class Main {
   public static void main(String[] args) throws IOException {
     CaesarInteractor caesarInteractor = CaesarInteractor.getInstance();
@@ -39,7 +38,7 @@ public class Main {
                              
                              """);
           String expression = reader.readLine().strip();
-          int res = polishNotation.evaluateExpression(expression);
+          double res = polishNotation.evaluateExpression(expression);
           System.out.println("Answer: " + res);
           break;
         case "3" :
